feat(graph): add persistent graph views with network selections
- Introduce persistent storage for graph views in Postgres, including network selections and state.
- Add GraphViewModel SQLAlchemy model and corresponding migration.
- Refactor ViewService to use database for listing, creating, retrieving, and deleting views.
- Update API endpoints to support new view persistence and detailed view info.
- Enhance frontend with network view saving, loading, and exporting capabilities.
- Remove deprecated saved view UI components and replace with new network view dialog.
- Add comprehensive React components for network pane, view page, export menu, and save dialog.

Test plan:
- Verified API endpoints for graph views with unit and integration tests.
- Tested network view saving, loading, and exporting in UI.
- Manual testing of graph network interactions and view persistence.
- Ensured no regressions in graph filtering and network display.
