+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/components/graph/graph-section-nav.tsx
@@ -14,8 +14,8 @@ const SECTIONS: Array<{
   path: string;
 }> = [
   { id: 'network', label: 'Network', icon: Network, path: 'network' },
-  { id: 'explore', label: 'Explore', icon: Search, path: 'explore' },
   { id: 'individuals', label: 'Individuals', icon: Users, path: 'individuals' },
+  { id: 'explore', label: 'Explore', icon: Search, path: 'explore' },
 ];
 
 export function GraphSectionNav({
@@ -23,13 +23,18 @@ export function GraphSectionNav({
   active,
   trailing,
   onNavigate,
+  sections,
 }: {
   workspaceId: string;
   active: GraphSection;
   trailing?: ReactNode;
   onNavigate?: (section: GraphSection) => void;
+  sections?: GraphSection[];
 }) {
   const router = useRouter();
+  const visibleSections = sections
+    ? SECTIONS.filter((section) => sections.includes(section.id))
+    : SECTIONS;
 
   const handleNavigate = (section: GraphSection, path: string) => {
     if (onNavigate) {
@@ -42,7 +47,7 @@ export function GraphSectionNav({
   return (
     <div className="flex h-10 items-center justify-between border-b bg-muted/30 px-4">
       <div className="flex items-center gap-1">
-        {SECTIONS.map(({ id, label, icon: Icon, path }) => (
+        {visibleSections.map(({ id, label, icon: Icon, path }) => (
           <button
             key={id}
             type="button"
